+++ src/main/groovy/life/qbic/datamodel/datasets/datastructure/files/FastqFile.groovy
@@ -1,4 +1,4 @@
-package life.qbic.datamodel.datasets.datastructure
+package life.qbic.datamodel.datasets.datastructure.files
 
 /**
  * A specialisation of a DataFile, represents an NGS fastq file
